backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Seina Tohyama <tiara818seina@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.0.0"
autopep8 = "^1.5.7"
python-dotenv = "^0.17.1"
Flask-Cors = "^3.0.10"
Flask-RESTful = "^0.3.9"
SQLAlchemy = "^1.4.16"
Flask-SQLAlchemy = "^2.5.1"
Flask-Migrate = "^3.0.0"
flask-marshmallow = "^0.14.0"
marshmallow-sqlalchemy = "^0.26.0"
PyMySQL = "^1.0.2"
SQLAlchemy-Utils = "^0.37.4"
cryptography = "^3.4.7"
Flask-Session = "^0.3.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
